Section_091_-_Day_91_-_Professional_Portfolio_Project_-_[HTTP_Requests_&_APIs]/pdf-to-speech/requirements.txt
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
Flask==3.1.3
gTTS==2.5.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.6.3
Werkzeug==3.1.6
pypdf
mutagen